A bunch of improvements
- make db respect main thread
- use async db calls
- add fab for adding routines

+++ app/src/main/java/de/janhopp/gymbro/model/workout/WorkoutRoutine.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class WorkoutRoutine(
-    @PrimaryKey(autoGenerate = true) val id: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int,
     val name: String,
     val description: String?,
     val emoji: String?,
