+++ app/src/main/java/de/janhopp/gymbro/ui/navigation/Destinations.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 sealed interface Destination {
     @Serializable data object WorkoutRoutinePicker : Destination
-    @Serializable data class WorkoutRoutineOverview(val routineId: Long) : Destination
+    @Serializable data class WorkoutRoutineOverview(val routineId: Int) : Destination
 }
